+++ changelog.md
@@ -5,6 +5,8 @@
 1. REMOVED unit tests redundant with e2e tests.
 1. MODIFIED release pipeline, splitting `try_publish` into `crates_publish` and `docker_publish`.
 1. MODIFIED `Dockerfile` to build from `cargo install` instead of the local source tree.
+1. REMOVED `osm_pbf_file` mod.
+1. ADDED `osm_pbf_file` domain.
 
 ## **2026-08-30** - 0.0.5